+++ src/domain/rum/RumCollection.ts
@@ -1,29 +1,34 @@
 import { EventManager } from '../../event';
 import type { FormatHooks } from '../../assembly';
 import { ErrorCollection, CrashCollection } from './error';
+import { OperationCollection } from './operation';
 import { ViewCollection } from './view';
 
 export class RumCollection {
   private constructor(
     private readonly viewCollection: ViewCollection,
-    private readonly errorCollection: ErrorCollection
+    private readonly errorCollection: ErrorCollection,
+    private readonly operationCollection: OperationCollection
   ) {}
 
   static async start(eventManager: EventManager, hooks: FormatHooks): Promise<RumCollection> {
     const viewCollection = await ViewCollection.start(eventManager, hooks);
     const errorCollection = new ErrorCollection(eventManager);
+    const operationCollection = new OperationCollection(eventManager);
     CrashCollection.start(eventManager);
-    return new RumCollection(viewCollection, errorCollection);
+    return new RumCollection(viewCollection, errorCollection, operationCollection);
   }
 
   getApi() {
     return {
       ...this.errorCollection.getApi(),
+      ...this.operationCollection.getApi(),
     };
   }
 
   stop(): void {
     this.viewCollection.stop();
     this.errorCollection.stop();
+    this.operationCollection.stop();
   }
 }