+++ src/domain/rum/index.ts
@@ -1,4 +1,5 @@
 export * from './RumCollection';
-export { ErrorOptions } from './error';
+export type { ErrorOptions } from './error';
+export type { FailureReason, FeatureOperationOptions } from './operation';
 export * from './rawRumData.types';
 export * from './rumEvent.types';